Refactoring and bug fixes
