+++ package.json
@@ -89,6 +89,7 @@
     "diff": "7.0.0",
     "franc-min": "6.2.0",
     "get-xpath": "3.3.0",
+    "js-beautify": "^1.15.4",
     "jsdom": "26.1.0",
     "urijs": "1.19.11"
   },
